Introduce static AssetLoader
